Add BSD 3-clause.
I'm hoping to follow a similar model to Sentry[1] with the open
sourcing, in that there exists a hosted (paid for) solution, but you can
run it yourself if you want. That said, unless you're a contributor
either by code or finance, I can't promise I have the time or energy to
answer any support queries.

[1] https://github.com/getsentry/sentry