Remove legacy input methods.
"County Team" based entry used for one event only, and the "mobile"
version of arrow-by-arrow entry that stopped working some time ago.

It would be good to redo the admin-based arrow input at some point using
React, and make sure it works properly on mobile, but it's not likely to
be the first priority.