Show all the categories when scoring, make score sheet button smaller
Definitely running into some duplication issues with the APIs now. I'm
not sure if I want to be using DRF Serializers, or just custom helper
functions? The Serializer will become more obvious if I had bigger
forms, but there's only really two or three proper forms (archer/entry
create and competition edit), neither of which are immediate on the
roadmap. When we're only reading data, and we're slightly mangling it to
make it appear flatter in the JS, helper functions are likely easier.