Use Map() objects rather than index finding and splicing.
Required some polyfilling for MapIterator.map(), so I needed to add
core-js and update the babel config method.

I'm not sure this is actually more effecient for bosses when thinking
about how we manipulate them, as you can't .splice() or .insert() in a
Map(). However, on things like targetList.sessions and
unallocatedEntries it seems much nicer, being able to instant lookup by
id to add or remove that entry.