Django 5 (#127)
* Update version of django and python.

Floppyforms is no longer compatible and frankly should be dead, so let's
kill it

* Remove floppyforms.

It's really a redundant package now that django has better form
rendering. The {% formrow %} tag is a bit of a loss, but it's not that
bad to type out custom layouts. Kept css changes to a minimum for now.

* Of course python is 3?

* wtf

* Looks like I had two venv names??

Also break caches on pips

* Fix timezones in tests