Refactor TargetList for useImmerReducer (#126)
* Refactor TargetList to use an Immer based reducer.

For now, it doesn't save anything, but saves actions in
targetList.actionsToSave. The next step will be to process these in the
TargetListProvider.

* Refactor the saving to use an action queue (and no retries)

* Remove <View> in favour of useQuery.

Also add a loading spinner.

* Move query saving into useActionQueue

* See if I can easily add eslint to circle

* Vendor useMousetrap (it's very small anyway)

* Use Map() objects rather than index finding and splicing.

Required some polyfilling for MapIterator.map(), so I needed to add
core-js and update the babel config method.

I'm not sure this is actually more effecient for bosses when thinking
about how we manipulate them, as you can't .splice() or .insert() in a
Map(). However, on things like targetList.sessions and
unallocatedEntries it seems much nicer, being able to instant lookup by
id to add or remove that entry.

* requestId should use a Ref

* useActionQueue should use a Reducer

* Send multiple actions at once.

* Make retries work in useActionQueue.

Currently infinitely back off retries with 1 increments.

* Add a Doctype!!

* Add an error boundary.

At least this is better than failing silently.

* Use ErrorBoundary, CompetitionContext in targetInput

* Convert input scores to reducers and queues

* Fix SHIFTUP action